move tests for 1:1 correspondence
